refactor ooxml decryption
