+++ crypto/CMakeLists.txt
@@ -1,5 +1,4 @@
 add_library(odr_crypto STATIC
-        src/CfbCrypto.cpp
         src/CryptoUtil.cpp
         )
 target_include_directories(odr_crypto PUBLIC include)
